# docs/en/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Speaking Coach API
  description: REST API for the IELTS Speaking Assistant. Provides voice-based conversation
    practice, pronunciation assessment, grammar feedback, and spaced-repetition flashcards.
    All endpoints except `/health` require a Bearer JWT obtained via POST /api/auth/login.
  version: 1.0.0
paths:
...
  /api/chat/transcribe:
    post:
      tags:
      - chat
      summary: Transcribe an uploaded audio file to text (STT only, no LLM)
      description: 'Lightweight STT-only endpoint: transcribes audio via Groq Whisper.

        No LLM, no TTS, no database writes — optimised for low-latency transcription.'
      operationId: transcribe_audio_only_api_chat_transcribe_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_transcribe_audio_only_api_chat_transcribe_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/chat/respond:
    post:
      tags:
      - chat
      summary: Send a message and receive an AI response with optional TTS audio
      description: 'Send a message (text or audio) and get an AI coach reply with
        TTS audio.


        Audio input is transcribed via STT when no text is provided. Input and output

        pass through guardrails. A new conversation is created if conversation_id
        is

        omitted; category is matched to a topics row to enforce the 5-session limit.

        Returns the reply text, inline or URL-referenced MP3, and a grammar summary.'
      operationId: chat_respond_api_chat_respond_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_chat_respond_api_chat_respond_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/audio/{storage_key}:
    get:
      tags:
      - audio
      summary: Stream a stored audio file from MinIO (requires auth, 1-hour cache)
      description: 'Stable audio proxy — streams a MinIO object to the client.

        Uses a stable URL (not a presigned URL) so the browser can cache the response.

        Auth is required.'
      operationId: stream_audio_api_audio__storage_key__get
      security:
      - HTTPBearer: []
      parameters:
      - name: storage_key
        in: path
        required: true
        schema:
          type: string
          title: Storage Key
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_chat_respond_api_chat_respond_post:
      properties:
        text:
          anyOf:
          - type: string
          - type: 'null'
          title: Text
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        topic:
          anyOf:
          - type: string
          - type: 'null'
          title: Topic
        voice_gender:
          anyOf:
          - type: string
          - type: 'null'
          title: Voice Gender
        voice_accent:
          anyOf:
          - type: string
          - type: 'null'
          title: Voice Accent
        audio_file:
          anyOf:
          - type: string
            contentMediaType: application/octet-stream
          - type: 'null'
          title: Audio File
        conversation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Conversation Id
      type: object
      title: Body_chat_respond_api_chat_respond_post
...
    Body_transcribe_audio_only_api_chat_transcribe_post:
      properties:
        audio_file:
          type: string
          contentMediaType: application/octet-stream
          title: Audio File
      type: object
      required:
      - audio_file
      title: Body_transcribe_audio_only_api_chat_transcribe_post
...
    ChatResponse:
      properties:
        user_input:
          type: string
          title: User Input
        response_text:
          type: string
          title: Response Text
        audio_base64:
          type: string
          title: Audio Base64
          default: ''
        audio_mime:
          type: string
          title: Audio Mime
          default: audio/mpeg
        user_audio_url:
          anyOf:
          - type: string
          - type: 'null'
          title: User Audio Url
        assistant_audio_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Assistant Audio Url
        conversation_id:
          type: string
          title: Conversation Id
        user_message_id:
          anyOf:
          - type: string
          - type: 'null'
          title: User Message Id
        grammar_summary:
          $ref: '#/components/schemas/GrammarSummary'
        grammar_detail:
          anyOf:
          - $ref: '#/components/schemas/GrammarDetailResponse'
          - type: 'null'
        tool_steps:
          items:
            $ref: '#/components/schemas/ToolCallStep'
          type: array
          title: Tool Steps
        suggestions:
          items:
            type: string
          type: array
          title: Suggestions
      type: object
      required:
      - user_input
      - response_text
      - conversation_id
      title: ChatResponse
...
    GrammarDetailResponse:
      properties:
        message_id:
          type: string
          title: Message Id
        user_input:
          type: string
          title: User Input
        errors:
          items:
            $ref: '#/components/schemas/GrammarErrorDetail'
          type: array
          title: Errors
        corrected_sentence:
          anyOf:
          - type: string
          - type: 'null'
          title: Corrected Sentence
        overall_score:
          type: integer
          title: Overall Score
      type: object
      required:
      - message_id
      - user_input
      - errors
      - corrected_sentence
      - overall_score
      title: GrammarDetailResponse
    GrammarErrorDetail:
      properties:
        id:
          type: integer
          title: Id
        original:
          type: string
          title: Original
        corrected:
          type: string
          title: Corrected
        start_char:
          type: integer
          title: Start Char
        end_char:
          type: integer
          title: End Char
        category:
          type: string
          title: Category
        severity:
          type: string
          title: Severity
        explanation:
          type: string
          title: Explanation
        rule:
          anyOf:
          - type: string
          - type: 'null'
          title: Rule
        example:
          anyOf:
          - type: string
          - type: 'null'
          title: Example
      type: object
      required:
      - id
      - original
      - corrected
      - start_char
      - end_char
      - category
      - severity
      - explanation
      title: GrammarErrorDetail
    GrammarSpan:
      properties:
        original:
          type: string
          title: Original
        corrected:
          type: string
          title: Corrected
        start_char:
          type: integer
          title: Start Char
        end_char:
          type: integer
          title: End Char
      type: object
      required:
      - original
      - corrected
      - start_char
      - end_char
      title: GrammarSpan
    GrammarSummary:
      properties:
        error_count:
          type: integer
          title: Error Count
        has_errors:
          type: boolean
          title: Has Errors
        flagged_spans:
          items:
            $ref: '#/components/schemas/GrammarSpan'
          type: array
          title: Flagged Spans
      type: object
      required:
      - error_count
      - has_errors
      - flagged_spans
      title: GrammarSummary
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ToolCallStep:
      properties:
        tool_name:
          type: string
          title: Tool Name
        input_summary:
          type: string
          title: Input Summary
        output_summary:
          type: string
          title: Output Summary
        result_label:
          type: string
          title: Result Label
          default: ''
        duration_ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration Ms
        status:
          type: string
          enum:
          - completed
          - failed
          title: Status
          default: completed
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - tool_name
      - input_summary
      - output_summary
      title: ToolCallStep
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer